chore: Remove a redundant comment from a path exclude
A path exclude with "reason" EXAMPLE_OF tells that the directory
contains examples which are not distributed with the release
artifact. So, stop repeating that in a comment.

Furthermore, making conclusions about the applicability of licenses
within path excludes is not the right place. This should be done in
the policy rules implementation, because there may be licenses which
are problematic even though excluded.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>